Updated to mod_attforblock v2016113001 r3.2.2 block_attendance v2017020700 r3.2.1

+++ mod/attendance/sessions.php
@@ -59,7 +59,6 @@ $PAGE->set_url($att->url_sessions(array('action' => $pageparams->action)));
 $PAGE->set_title($course->shortname. ": ".$att->name);
 $PAGE->set_heading($course->fullname);
 $PAGE->set_cacheable(true);
-$PAGE->set_button($OUTPUT->update_module_button($cm->id, 'attendance'));
 $PAGE->navbar->add($att->name);
 
 $currenttab = attendance_tabs::TAB_ADD;
@@ -113,9 +112,7 @@ switch ($att->pageparams->action) {
 
         if (isset($confirm) && confirm_sesskey()) {
             $att->delete_sessions(array($sessionid));
-            if ($att->grade > 0) {
-                attendance_update_all_users_grades($att, $cm);
-            }
+            attendance_update_users_grade($att);
             redirect($att->url_manage(), get_string('sessiondeleted', 'attendance'));
         }
 
@@ -136,16 +133,16 @@ switch ($att->pageparams->action) {
         exit;
     case mod_attendance_sessions_page_params::ACTION_DELETE_SELECTED:
         $confirm    = optional_param('confirm', null, PARAM_INT);
+        $message = get_string('deletecheckfull', '', get_string('session', 'attendance'));
 
         if (isset($confirm) && confirm_sesskey()) {
             $sessionsids = required_param('sessionsids', PARAM_ALPHANUMEXT);
             $sessionsids = explode('_', $sessionsids);
-
-            $att->delete_sessions($sessionsids);
-            if ($att->grade > 0) {
-                attendance_update_all_users_grades($att, $cm);
+            if ($att->pageparams->action == mod_attendance_sessions_page_params::ACTION_DELETE_SELECTED) {
+                $att->delete_sessions($sessionsids);
+                attendance_update_users_grade($att);
+                redirect($att->url_manage(), get_string('sessiondeleted', 'attendance'));
             }
-            redirect($att->url_manage(), get_string('sessiondeleted', 'attendance'));
         }
         $sessid = optional_param_array('sessid', '', PARAM_SEQUENCE);
         if (empty($sessid)) {
@@ -153,7 +150,6 @@ switch ($att->pageparams->action) {
         }
         $sessionsinfo = $att->get_sessions_info($sessid);
 
-        $message = get_string('deletecheckfull', '', get_string('session', 'attendance'));
         $message .= html_writer::empty_tag('br');
         foreach ($sessionsinfo as $sessinfo) {
             $message .= html_writer::empty_tag('br');
@@ -236,6 +232,10 @@ function construct_sessions_data_for_add($formdata) {
     $sessiondate = $formdata->sessiondate + $sesstarttime;
     $duration = $sesendtime - $sesstarttime;
     $now = time();
+
+    if (empty(get_config('attendance', 'studentscanmark'))) {
+        $formdata->studentscanmark = 0;
+    }
 
     $sessions = array();
     if (isset($formdata->addmultiply)) {
